Incorporate the testify library.
After long hesitation, I, with the blessing of @abhinav decided to
incorporate the testify library to make it simpler and easier to hack it
for our needs, like improving stacktraces in our specific case, and/or
other test messages.

We might want to keep the library API and behaviors as close to the
original as possible, so that one day we might even be able to send
back patches upstream.